requirements.txt
# ---------------- ENV + ASYNC ----------------
python-dotenv==1.0.1
asyncio==3.4.3

# ---------------- LANGCHAIN ECOSYSTEM ----------------
langchain
langchain-core
langchain-community
langchain-groq

# ---------------- LANGGRAPH ----------------
langgraph
langgraph-sdk

# ---------------- MCP SUPPORT ----------------
langchain-mcp-adapters==0.1.14
mcp>=1.9.2

# ---------------- NETWORKING ----------------
requests==2.32.3
aiohttp==3.10.5
httpx==0.27.2

# ---------------- UTILS ----------------
pydantic==2.12.5
rich==13.9.3
loguru==0.7.2

# ---------------- FASTAPI BACKEND ----------------
fastapi==0.111.0
uvicorn[standard]==0.30.1
